Extracao de endereco do cliente dto, pedidos recuperados no getall
